modify: show sessions on top page
