+++ lib/pages/index.dart
@@ -4,6 +4,7 @@ import 'package:confwebsite2022/widgets/background.dart';
 import 'package:confwebsite2022/widgets/custom_button.dart';
 import 'package:confwebsite2022/widgets/features.dart';
 import 'package:confwebsite2022/widgets/footer.dart';
+import 'package:confwebsite2022/widgets/sessions.dart';
 import 'package:confwebsite2022/widgets/social.dart';
 import 'package:confwebsite2022/widgets/staff.dart';
 import 'package:flutter/material.dart';
@@ -222,103 +223,82 @@ class Body extends StatelessWidget {
       final appLocalizations = AppLocalizations.of(context)!;
       final sizeFactor = (layout == ResponsiveLayout.slim) ? 0.6 : 1.0;
 
-      return CustomScrollView(
+      return ListView(
         shrinkWrap: true,
-        slivers: [
-          SliverFillRemaining(
-            hasScrollBody: false,
-            child: Column(
-              children: [
-                Container(
-                  padding: const EdgeInsets.symmetric(vertical: 16),
-                  child: Column(
-                    children: [
-                      SvgPicture.asset(
-                        Assets.flutterkaigiLogo,
-                        width: logoWidth * sizeFactor,
-                      ),
-                      FittedBox(
-                        fit: BoxFit.contain,
-                        child: Text(
-                          'FlutterKaigi',
-                          style:
-                              titleTextStyle.apply(fontSizeFactor: sizeFactor),
-                        ),
-                      ),
-                      Gap(32 * sizeFactor),
-                      FittedBox(
-                        fit: BoxFit.contain,
-                        child: Text(
-                          '@ONLINE / November 16-18, 2022',
-                          style: subtitleTextStyle.apply(
-                              fontSizeFactor: sizeFactor),
-                        ),
-                      ),
-                      const Gap(32),
-                      CustomButton(
-                        isShow: initialLaunched,
-                        colors: initialLaunched
-                            ? const [Colors.blue, Colors.green]
-                            : null,
-                        title: appLocalizations.checkLatestNews,
-                        message: appLocalizations.tweet,
-                        onPress: () async {
-                          await launch(
-                            'https://twitter.com/FlutterKaigi',
-                            webOnlyWindowName: '_blank',
-                          );
-                        },
-                      ),
-                      const Gap(16),
-                      Center(
-                        child: Row(
-                          mainAxisAlignment: MainAxisAlignment.center,
-                          children: [
-                            CustomButton(
-                              isShow: true,
-                              colors: const [Colors.orange, Colors.red],
-                              title: appLocalizations.schedule,
-                              message: appLocalizations.showSchedule,
-                              onPress: () => launch(
-                                'https://fortee.jp/flutterkaigi-2022/timetable',
-                                webOnlyWindowName: '_blank',
-                              ),
-                            ),
-                            CustomButton(
-                              isShow: !initialLaunched,
-                              colors: announceSponsor
-                                  ? const [Colors.blue, Colors.indigo]
-                                  : null,
-                              title: appLocalizations.sponsor,
-                              message: announceSponsor
-                                  ? appLocalizations.becomeSponsor
-                                  : appLocalizations.waitFor,
-                              onPress: announceSponsor
-                                  ? () async {
-                                      await launch(
-                                        startSponsor
-                                            ? 'https://fortee.jp/flutterkaigi-2022/sponsor/form'
-                                            : 'https://docs.google.com/presentation/d/1HEwDIi6rxzKUnZmu7EKkwR04bvTQnSjWjpw3ldunczM/edit?usp=sharing',
-                                        webOnlyWindowName: '_blank',
-                                      );
-                                    }
-                                  : null,
-                            ),
-                          ],
-                        ),
-                      ),
-                      const Gap(16),
-                      const Social(),
-                    ],
-                  ),
-                ),
-                const Gap(32),
-                const StaffSection(),
-                const Spacer(),
-                Footer(layout: layout),
-              ],
+        padding: const EdgeInsets.only(top: 16),
+        children: [
+          SvgPicture.asset(
+            Assets.flutterkaigiLogo,
+            width: logoWidth * sizeFactor,
+          ),
+          Center(
+            child: Text(
+              'FlutterKaigi',
+              style: titleTextStyle.apply(fontSizeFactor: sizeFactor),
             ),
           ),
+          Gap(32 * sizeFactor),
+          Center(
+            child: Text(
+              '@ONLINE / November 16-18, 2022',
+              style: subtitleTextStyle.apply(fontSizeFactor: sizeFactor),
+            ),
+          ),
+          const Gap(32),
+          CustomButton(
+            isShow: initialLaunched,
+            colors: initialLaunched ? const [Colors.blue, Colors.green] : null,
+            title: appLocalizations.checkLatestNews,
+            message: appLocalizations.tweet,
+            onPress: () async {
+              await launch(
+                'https://twitter.com/FlutterKaigi',
+                webOnlyWindowName: '_blank',
+              );
+            },
+          ),
+          const Gap(16),
+          Row(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: [
+              CustomButton(
+                isShow: true,
+                colors: const [Colors.orange, Colors.red],
+                title: appLocalizations.schedule,
+                message: appLocalizations.showSchedule,
+                onPress: () => launch(
+                  'https://fortee.jp/flutterkaigi-2022/timetable',
+                  webOnlyWindowName: '_blank',
+                ),
+              ),
+              CustomButton(
+                isShow: !initialLaunched,
+                colors:
+                    announceSponsor ? const [Colors.blue, Colors.indigo] : null,
+                title: appLocalizations.sponsor,
+                message: announceSponsor
+                    ? appLocalizations.becomeSponsor
+                    : appLocalizations.waitFor,
+                onPress: announceSponsor
+                    ? () async {
+                        await launch(
+                          startSponsor
+                              ? 'https://fortee.jp/flutterkaigi-2022/sponsor/form'
+                              : 'https://docs.google.com/presentation/d/1HEwDIi6rxzKUnZmu7EKkwR04bvTQnSjWjpw3ldunczM/edit?usp=sharing',
+                          webOnlyWindowName: '_blank',
+                        );
+                      }
+                    : null,
+              ),
+            ],
+          ),
+          const Gap(16),
+          const Social(),
+          const Gap(32),
+          const SessionList(),
+          const Gap(32),
+          const StaffSection(),
+          Footer(layout: layout),
         ],
       );
     });
