+++ lib/router/router_provider.dart
@@ -1,5 +1,4 @@
 import 'package:confwebsite2022/pages/index.dart';
-import 'package:confwebsite2022/pages/session.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
@@ -11,15 +10,6 @@ final routerProvider = Provider(
       GoRoute(
         path: '/',
         builder: (_, __) => const TopPage(),
-        routes: [
-          GoRoute(
-            path: 'session',
-            pageBuilder: (_, state) => NoTransitionPage(
-              key: state.pageKey,
-              child: const SessionPage(),
-            ),
-          ),
-        ],
       ),
     ],
     errorPageBuilder: (context, state) => MaterialPage(
