+++ lib/widgets/footer.dart
@@ -5,7 +5,6 @@ import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:gap/gap.dart';
-import 'package:go_router/go_router.dart';
 import 'package:styled_text/styled_text.dart';
 import 'package:url_launcher/url_launcher.dart';
 
@@ -42,11 +41,6 @@ class Footer extends StatelessWidget {
           text: item['name'],
           onPressed: () => launch(item['url']!),
         ),
-      _FooterButton(
-        message: appLocalizations.session,
-        text: appLocalizations.session,
-        onPressed: () => context.go('/session'),
-      ),
       _FooterButton(
         message: appLocalizations.licenses,
         text: appLocalizations.licenses,
